@@ -3,12 +3,12 @@
  * @author shiloh
  * @date 2024/7/22 9:12
  */
-import { boot } from 'quasar/wrappers'
+import { defineBoot } from '#q-app/wrappers'
 import axios from 'axios'
 
 const api = axios.create({ baseURL: 'https://jsonplaceholder.typicode.com' })
 
-export default boot(({ app }) => {
+export default defineBoot(({ app }) => {
   // for use inside Vue files (Options API) through this.$axios and this.$api
 
   app.config.globalProperties.$axios = axios
