@@ -1,14 +1,6 @@
 <template>
-  <q-item
-    clickable
-    tag="a"
-    target="_blank"
-    :href="link"
-  >
-    <q-item-section
-      v-if="icon"
-      avatar
-    >
+  <q-item clickable tag="a" target="_blank" :href="link">
+    <q-item-section v-if="icon" avatar>
       <q-icon :name="icon" />
     </q-item-section>
 
@@ -20,20 +12,20 @@
 </template>
 
 <script setup lang="ts">
-  defineOptions({
-    name: 'EssentialLink'
-  })
+defineOptions({
+  name: 'EssentialLink',
+})
 
-  export interface EssentialLinkProps {
-    title?: string
-    caption?: string
-    link?: string
-    icon?: string
-  }
+export interface EssentialLinkProps {
+  title?: string
+  caption?: string
+  link?: string
+  icon?: string
+}
 
-  withDefaults(defineProps<EssentialLinkProps>(), {
-    caption: '',
-    link: '#',
-    icon: ''
-  })
+withDefaults(defineProps<EssentialLinkProps>(), {
+  caption: '',
+  link: '#',
+  icon: '',
+})
 </script>
