@@ -2,11 +2,7 @@
   <div>
     <p>{{ title }}</p>
     <ul>
-      <li
-        v-for="todo in todos"
-        :key="todo.id"
-        @click="increment"
-      >
+      <li v-for="todo in todos" :key="todo.id" @click="increment">
         {{ todo.id }} - {{ todo.content }}
       </li>
     </ul>
@@ -17,26 +13,26 @@
 </template>
 
 <script setup lang="ts">
-  import { computed, ref } from 'vue'
-  import type { Meta, Todo } from './models'
+import { computed, ref } from 'vue'
+import type { Meta, Todo } from './models'
 
-  interface Props {
-    title: string
-    todos?: Todo[]
-    meta: Meta
-    active: boolean
-  }
+interface Props {
+  title: string
+  todos?: Todo[]
+  meta: Meta
+  active: boolean
+}
 
-  const props = withDefaults(defineProps<Props>(), {
-    todos: () => []
-  })
+const props = withDefaults(defineProps<Props>(), {
+  todos: () => [],
+})
 
-  const clickCount = ref(0)
+const clickCount = ref(0)
 
-  function increment() {
-    clickCount.value += 1
-    return clickCount.value
-  }
+function increment() {
+  clickCount.value += 1
+  return clickCount.value
+}
 
-  const todoCount = computed(() => props.todos.length)
+const todoCount = computed(() => props.todos.length)
 </script>
