@@ -19,8 +19,8 @@
 </template>
 
 <script setup lang="ts">
-  import HeaderView from 'layouts/components/HeaderView.vue'
-  import SidebarMenu from 'layouts/components/SidebarMenu.vue'
-  import FooterView from 'layouts/components/FooterView.vue'
-  import MainContent from 'layouts/components/MainContent.vue'
+import HeaderView from 'layouts/components/HeaderView.vue'
+import SidebarMenu from 'layouts/components/SidebarMenu.vue'
+import FooterView from 'layouts/components/FooterView.vue'
+import MainContent from 'layouts/components/MainContent.vue'
 </script>
