@@ -5,11 +5,7 @@
 -->
 <template>
   <q-breadcrumbs active-color="text-white">
-    <q-breadcrumbs-el
-      :label="HOME_MENU.label"
-      :icon="HOME_MENU.icon"
-      :to="HOME_MENU.path"
-    />
+    <q-breadcrumbs-el :label="HOME_MENU.label" :icon="HOME_MENU.icon" :to="HOME_MENU.path" />
     <q-breadcrumbs-el
       v-for="breadcrumb in breadcrumbs"
       :key="breadcrumb.label"
@@ -20,45 +16,45 @@
 </template>
 
 <script setup lang="ts">
-  import { onMounted, ref, watch } from 'vue'
-  import { useRouter } from 'vue-router'
-  import { HOME_MENU } from 'src/router/routes/menu.data'
+import { onMounted, ref, watch } from 'vue'
+import { useRouter } from 'vue-router'
+import { HOME_MENU } from 'src/router/routes/menu.data'
 
-  defineOptions({ name: 'BreadcrumbView' })
+defineOptions({ name: 'BreadcrumbView' })
 
-  export interface Breadcrumb {
-    icon?: string
-    label: string
-    to: string
-  }
+export interface Breadcrumb {
+  icon?: string
+  label: string
+  to: string
+}
 
-  const breadcrumbs = ref<Breadcrumb[]>([])
+const breadcrumbs = ref<Breadcrumb[]>([])
 
-  const router = useRouter()
-  const { currentRoute } = router
-  onMounted(() => {
+const router = useRouter()
+const { currentRoute } = router
+onMounted(() => {
+  createBreadcrumbs()
+})
+
+watch(
+  () => currentRoute.value,
+  () => {
     createBreadcrumbs()
-  })
+  },
+)
 
-  watch(
-    () => currentRoute.value,
-    () => {
-      createBreadcrumbs()
-    }
-  )
-
-  function createBreadcrumbs() {
-    breadcrumbs.value =
-      HOME_MENU.path === currentRoute.value.path
-        ? []
-        : [
-            {
-              label: currentRoute.value.name as string,
-              to: currentRoute.value.path,
-              icon: currentRoute.value.meta.icon as string
-            }
-          ]
-  }
+function createBreadcrumbs() {
+  breadcrumbs.value =
+    HOME_MENU.path === currentRoute.value.path
+      ? []
+      : [
+          {
+            label: currentRoute.value.name as string,
+            to: currentRoute.value.path,
+            icon: currentRoute.value.meta.icon as string,
+          },
+        ]
+}
 </script>
 
 <style scoped></style>
