@@ -4,28 +4,17 @@
   * @date 2024/7/16 17:18
 -->
 <template>
-  <q-footer
-    elevated
-    class="bg-grey-8 text-white"
-    reveal
-  >
+  <q-footer elevated class="bg-grey-8 text-white" reveal>
     <q-toolbar>
       <q-toolbar-title class="text-center">
-        <q-btn
-          flat
-          dense
-          no-caps
-          target="_blank"
-          href="#"
-          label="Your Copyright © 2024"
-        />
+        <q-btn flat dense no-caps target="_blank" href="#" label="Your Copyright © 2024" />
       </q-toolbar-title>
     </q-toolbar>
   </q-footer>
 </template>
 
 <script setup lang="ts">
-  defineOptions({ name: 'FooterView' })
+defineOptions({ name: 'FooterView' })
 </script>
 
 <style scoped></style>
