@@ -4,20 +4,9 @@
   * @date 2024/7/16 17:14
 -->
 <template>
-  <q-header
-    elevated
-    class="bg-primary text-white"
-    height-hint="98"
-  >
+  <q-header elevated class="bg-primary text-white" height-hint="98">
     <q-toolbar>
-      <q-btn
-        dense
-        flat
-        round
-        class="q-mr-md"
-        icon="menu"
-        @click="sidebarStore.toggleCollapsed()"
-      />
+      <q-btn dense flat round class="q-mr-md" icon="menu" @click="sidebarStore.toggleCollapsed()" />
       <breadcrumb-view />
 
       <q-toolbar-title />
@@ -32,12 +21,12 @@
 </template>
 
 <script setup lang="ts">
-  import { useSidebarStore } from 'stores/sidebar'
-  import BreadcrumbView from 'layouts/components/BreadcrumbView.vue'
+import { useSidebarStore } from 'stores/sidebar'
+import BreadcrumbView from 'layouts/components/BreadcrumbView.vue'
 
-  const sidebarStore = useSidebarStore()
+const sidebarStore = useSidebarStore()
 
-  defineOptions({ name: 'HeaderView' })
+defineOptions({ name: 'HeaderView' })
 </script>
 
 <style scoped></style>
