@@ -6,19 +6,13 @@
 <template>
   <q-page-container class="bg-grey-2">
     <tab-view />
-    <q-page
-      class="row"
-      :style-fn="(offset) => ({ minHeight: `calc(100vh - ${offset + 40}px)` })"
-    >
+    <q-page class="row" :style-fn="(offset) => ({ minHeight: `calc(100vh - ${offset + 40}px)` })">
       <div class="col-12">
         <div class="column full-height">
           <q-scroll-area class="col bg-white q-ma-md shadow-2 rounded-borders q-pa-md">
             <router-view>
               <template #default="{ Component }">
-                <transition
-                  mode="in-out"
-                  :name="transitionName"
-                >
+                <transition mode="out-in" :name="transitionName">
                   <component :is="Component" />
                 </transition>
               </template>
@@ -31,29 +25,29 @@
 </template>
 
 <script setup lang="ts">
-  import { onMounted, ref } from 'vue'
-  import { useTabStore } from 'stores/tab'
-  import { useRoute } from 'vue-router'
-  import { MenuType } from 'src/router/routes/menu.data'
-  import TabView from 'layouts/components/TabView.vue'
+import { onMounted, ref } from 'vue'
+import { useTabStore } from 'stores/tab'
+import { useRoute } from 'vue-router'
+import { MenuType } from 'src/router/routes/menu.data'
+import TabView from 'layouts/components/TabView.vue'
 
-  defineOptions({ name: 'MainContent' })
-  const transitionName = ref('scale')
-  const tabStore = useTabStore()
-  const route = useRoute()
+defineOptions({ name: 'MainContent' })
+const transitionName = ref('scale')
+const tabStore = useTabStore()
+const route = useRoute()
 
-  onMounted(() => {
-    tabStore.addTab({
-      path: route.path,
-      label: route.name as string,
-      icon: route.meta.icon as string,
-      type: MenuType.PAGE
-    })
+onMounted(() => {
+  tabStore.addTab({
+    path: route.path,
+    label: route.name as string,
+    icon: route.meta.icon as string,
+    type: MenuType.PAGE,
   })
+})
 </script>
 
 <style scoped>
-  /*.fade-enter-active,
+/*.fade-enter-active,
   .fade-leave-active {
     transition: opacity 0.5s ease;
   }
@@ -88,18 +82,18 @@
     left: 0;
   }*/
 
-  .scale-enter-active,
-  .scale-leave-active {
-    transition: all 0.5s ease;
-  }
+.scale-enter-active,
+.scale-leave-active {
+  transition: all 0.5s ease;
+}
 
-  .scale-enter-from,
-  .scale-leave-to {
-    opacity: 0;
-    transform: scale(0.9);
-  }
+.scale-enter-from,
+.scale-leave-to {
+  opacity: 0;
+  transform: scale(0.9);
+}
 
-  /*.scale-slide-enter-active,
+/*.scale-slide-enter-active,
   .scale-slide-leave-active {
     position: absolute;
     transition: all 0.85s ease;
