@@ -4,13 +4,7 @@
   * @date 2024/7/16 17:16
 -->
 <template>
-  <q-drawer
-    show-if-above
-    v-model="sidebarStore.collapsed"
-    side="left"
-    bordered
-    :breakpoint="500"
-  >
+  <q-drawer show-if-above v-model="sidebarStore.collapsed" side="left" bordered :breakpoint="500">
     <q-scroll-area style="height: calc(100% - 70px); margin-top: 70px">
       <q-list>
         <q-item
@@ -28,10 +22,7 @@
           <q-item-section>{{ HOME_MENU.label }}</q-item-section>
         </q-item>
 
-        <template
-          v-for="menuItem in menuList"
-          :key="menuItem.label"
-        >
+        <template v-for="menuItem in menuList" :key="menuItem.label">
           <sidebar-menu-item :menu="menuItem" />
         </template>
       </q-list>
@@ -41,14 +32,8 @@
       class="absolute-top q-pa-sm flex items-center justify-center"
       style="height: 70px; gap: 10px"
     >
-      <q-avatar
-        class="q-mb-sm"
-        color="white"
-      >
-        <img
-          src="~assets/logo.svg"
-          alt="App Logo"
-        />
+      <q-avatar class="q-mb-sm" color="white">
+        <img src="~assets/logo.svg" alt="App Logo" />
       </q-avatar>
       <span class="text-weight-bold">Your System Name</span>
     </div>
@@ -56,18 +41,18 @@
 </template>
 
 <script setup lang="ts">
-  import { HOME_MENU, menuList } from 'src/router/routes/menu.data'
-  import SidebarMenuItem from 'layouts/components/SidebarMenuItem.vue'
-  import { useSidebarStore } from 'stores/sidebar'
-  import { useRoute } from 'vue-router'
-  import { useTabStore } from 'stores/tab'
+import { HOME_MENU, menuList } from 'src/router/routes/menu.data'
+import SidebarMenuItem from 'layouts/components/SidebarMenuItem.vue'
+import { useSidebarStore } from 'stores/sidebar'
+import { useRoute } from 'vue-router'
+import { useTabStore } from 'stores/tab'
 
-  defineOptions({ name: 'SidebarMenu' })
+defineOptions({ name: 'SidebarMenu' })
 
-  const sidebarStore = useSidebarStore()
-  const route = useRoute()
+const sidebarStore = useSidebarStore()
+const route = useRoute()
 
-  const tabStore = useTabStore()
+const tabStore = useTabStore()
 </script>
 
 <style scoped></style>
