@@ -1,5 +1,6 @@
 import type { RouteRecordRaw } from 'vue-router'
-import { HOME_MENU, Menu, menuList, MenuType } from 'src/router/routes/menu.data'
+import type { Menu } from 'src/router/routes/menu.data'
+import { HOME_MENU, menuList, MenuType } from 'src/router/routes/menu.data'
 // Vite 一次性导入匹配指定路径的所有组件实例，返回值为一个 key = 路径，value = 组件实例 的对象
 // 在 Vue3 中，如果需要根据菜单配置或者后台接口数据生成动态路由，则要先将对应的组件全部导入进来
 // 然后通过菜单的某个字段值（如：component）去匹配
@@ -15,10 +16,10 @@ const home: RouteRecordRaw = {
       name: HOME_MENU.name,
       component: () => import('src/views/IndexPage.vue'),
       meta: {
-        icon: HOME_MENU.icon
-      }
-    }
-  ]
+        icon: HOME_MENU.icon,
+      },
+    },
+  ],
 }
 
 const dynamicRoutes: RouteRecordRaw[] = []
@@ -47,8 +48,8 @@ function generateDynamicRoutes(menus: Menu[]) {
         name: menu.label ?? menu.path,
         component: dynamicRoutesMap[`/src/views/${menu.component}.vue`],
         meta: {
-          icon: menu.icon
-        }
+          icon: menu.icon,
+        },
       } as RouteRecordRaw)
     }
   }
@@ -60,7 +61,7 @@ home.children.push({
   // Always leave this as last one,
   // but you can also remove it
   path: '/:catchAll(.*)*',
-  component: () => import('src/views/ErrorNotFound.vue')
+  component: () => import('src/views/ErrorNotFound.vue'),
 })
 
 export default [home]
