@@ -2,14 +2,14 @@ import { defineStore } from 'pinia'
 
 export const useCounterStore = defineStore('counter', {
   state: () => ({
-    counter: 0
+    counter: 0,
   }),
   getters: {
-    doubleCount: (state) => state.counter * 2
+    doubleCount: (state) => state.counter * 2,
   },
   actions: {
     increment() {
       this.counter++
-    }
-  }
+    },
+  },
 })
