@@ -1,5 +1,7 @@
-<template>UserProfile</template>
+<template>
+  <div>UserProfile</div>
+</template>
 
 <script setup lang="ts">
-  defineOptions({ name: 'ProfileView' })
+defineOptions({ name: 'ProfileView' })
 </script>
