@@ -3,10 +3,12 @@
   * @author shiloh
   * @date 2024/7/16 16:39
 -->
-<template>UserLogView</template>
+<template>
+  <div>UserLogView</div>
+</template>
 
 <script setup lang="ts">
-  defineOptions({ name: 'UserLogView' })
+defineOptions({ name: 'UserLogView' })
 </script>
 
 <style scoped></style>
