@@ -3,10 +3,10 @@
   * @author shiloh
   * @date 2024/7/16 15:31
 -->
-<template>RoleView</template>
+<template><div>RoleView</div></template>
 
 <script setup lang="ts">
-  defineOptions({ name: 'RoleView' })
+defineOptions({ name: 'RoleView' })
 </script>
 
 <style scoped></style>
