@@ -18,12 +18,7 @@
       :pagination="{ rowsPerPage: 10 }"
     >
       <template #top-left>
-        <q-btn
-          color="primary"
-          :disable="loading"
-          no-caps
-          label="Add User"
-        />
+        <q-btn color="primary" :disable="loading" no-caps label="Add User" />
         <q-btn
           class="q-ml-sm"
           color="red"
@@ -33,15 +28,12 @@
         />
       </template>
       <template #top-right>
-        <q-btn
-          color="primary"
-          icon-right="archive"
-          label="Export to csv"
-          no-caps
-        />
+        <q-btn color="primary" icon-right="archive" label="Export to csv" no-caps />
       </template>
       <template #body-cell-address="props">
-        <q-td :props="props"> {{ (props.row as User).address.city }} - {{ (props.row as User).address.street }} </q-td>
+        <q-td :props="props">
+          {{ (props.row as User).address.city }} - {{ (props.row as User).address.street }}
+        </q-td>
       </template>
       <template #body-cell-website="props">
         <q-td :props="props">
@@ -64,22 +56,8 @@
       </template>
       <template #body-cell-actions="props">
         <q-td :props="props">
-          <q-btn
-            color="primary"
-            flat
-            dense
-            no-caps
-            icon="visibility"
-          >
-          </q-btn>
-          <q-btn
-            color="primary"
-            flat
-            no-caps
-            icon="edit_square"
-            dense
-          >
-          </q-btn>
+          <q-btn color="primary" flat dense no-caps icon="visibility"> </q-btn>
+          <q-btn color="primary" flat no-caps icon="edit_square" dense> </q-btn>
           <q-btn
             color="red"
             flat
@@ -96,135 +74,137 @@
 </template>
 
 <script setup lang="ts">
-  import { QTableColumn, useQuasar } from 'quasar'
-  import { onMounted, ref } from 'vue'
-  import { api } from 'boot/axios'
-  import type { User } from './models'
+import type { QTableColumn } from 'quasar'
+import { useQuasar } from 'quasar'
+import { onMounted, ref } from 'vue'
+import { api } from 'boot/axios'
+import type { User } from './models'
 
-  defineOptions({ name: 'UserView' })
-  const $q = useQuasar()
+defineOptions({ name: 'UserView' })
+const $q = useQuasar()
 
-  // region user table data
+// region user table data
 
-  const columns: QTableColumn[] = [
-    {
-      name: 'id',
-      label: 'ID',
-      align: 'center',
-      field: 'id',
-      sortable: true
-    },
-    {
-      name: 'name',
-      label: 'Name',
-      align: 'center',
-      field: 'name'
-    },
-    {
-      name: 'username',
-      label: 'Username',
-      align: 'left',
-      field: 'username'
-    },
-    {
-      name: 'email',
-      label: 'Email',
-      align: 'right',
-      field: 'email'
-    },
-    {
-      name: 'address',
-      label: 'Address',
-      align: 'center',
-      field: 'address'
-    },
-    {
-      name: 'phone',
-      label: 'Phone',
-      align: 'center',
-      field: 'phone'
-    },
-    {
-      name: 'website',
-      label: 'Website',
-      align: 'left',
-      field: 'website'
-    },
-    {
-      name: 'company',
-      label: 'Company',
-      align: 'center',
-      field: 'company'
-    },
-    {
-      name: 'actions',
-      label: 'Actions',
-      align: 'center',
-      field: 'actions'
-    }
-  ]
-  const loading = ref(false)
-  const users = ref<User[]>([])
-  const selectedUsers = ref<User[]>([])
+const columns: QTableColumn[] = [
+  {
+    name: 'id',
+    label: 'ID',
+    align: 'center',
+    field: 'id',
+    sortable: true,
+  },
+  {
+    name: 'name',
+    label: 'Name',
+    align: 'center',
+    field: 'name',
+  },
+  {
+    name: 'username',
+    label: 'Username',
+    align: 'left',
+    field: 'username',
+  },
+  {
+    name: 'email',
+    label: 'Email',
+    align: 'right',
+    field: 'email',
+  },
+  {
+    name: 'address',
+    label: 'Address',
+    align: 'center',
+    field: 'address',
+  },
+  {
+    name: 'phone',
+    label: 'Phone',
+    align: 'center',
+    field: 'phone',
+  },
+  {
+    name: 'website',
+    label: 'Website',
+    align: 'left',
+    field: 'website',
+  },
+  {
+    name: 'company',
+    label: 'Company',
+    align: 'center',
+    field: 'company',
+  },
+  {
+    name: 'actions',
+    label: 'Actions',
+    align: 'center',
+    field: 'actions',
+  },
+]
+const loading = ref(false)
+const users = ref<User[]>([])
+const selectedUsers = ref<User[]>([])
 
-  // endregion
+// endregion
 
-  // region fetch users
+// region fetch users
 
-  const getUsers = async () => {
-    loading.value = true
-    try {
-      const res = await api.get('/users')
-      users.value = res.data as User[]
-      $q.notify({
-        type: 'positive',
-        position: 'top',
-        message: `Successfully obtained ${users.value.length} users!`,
-        icon: 'mood',
-        timeout: 3000
-      })
-    } finally {
-      loading.value = false
-    }
-  }
-
-  // endregion
-
-  // region delete
-
-  const handleDeleteUser = (user: User) => {
-    $q.dialog({
-      title: '<span class="text-warning">WARNING</span>',
-      message: `Are you sure you want to delete <span class="text-bold">${user.name}?</span>`,
-      cancel: true,
-      persistent: true,
-      html: true
+const getUsers = async () => {
+  loading.value = true
+  try {
+    const res = await api.get('/users')
+    users.value = res.data as User[]
+    $q.notify({
+      type: 'positive',
+      position: 'top',
+      message: `Successfully obtained ${users.value.length} users!`,
+      icon: 'mood',
+      timeout: 3000,
     })
-      .onOk(() => {
-        $q.notify({
-          color: 'positive',
-          message: 'User deleted!',
-          icon: 'check',
-          timeout: 3000,
-          position: 'top'
-        })
-      })
-      .onCancel(() => {
-        $q.notify({
-          color: 'negative',
-          message: 'User not deleted!',
-          icon: 'close',
-          timeout: 3000,
-          position: 'top'
-        })
-      })
+  } finally {
+    loading.value = false
   }
+}
 
-  // endregion
+// endregion
 
-  onMounted(async () => {
-    await getUsers()
+// region delete
+
+const handleDeleteUser = (user: User) => {
+  $q.dialog({
+    title: '<span class="text-warning">WARNING</span>',
+    message: `Are you sure you want to delete <span class="text-bold">${user.name}?</span>`,
+    cancel: true,
+    persistent: true,
+    html: true,
   })
+    .onOk(() => {
+      $q.notify({
+        color: 'positive',
+        message: 'User deleted!',
+        icon: 'check',
+        timeout: 3000,
+        position: 'top',
+      })
+    })
+    .onCancel(() => {
+      $q.notify({
+        color: 'negative',
+        message: 'User not deleted!',
+        icon: 'close',
+        timeout: 3000,
+        position: 'top',
+      })
+    })
+}
+
+// endregion
+
+onMounted(async () => {
+  console.log('$q', $q)
+  await getUsers()
+})
 </script>
 
 <style scoped></style>
